Satisfy warm restart CI gates
